pseudo_labeling/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
openmim
mmengine>=0.7.0
mmcv>=2.0.0
mmdet>=3.0.0
numpy
opencv-python
pillow
pycocotools
tqdm

